refactor: adjust attributes to be alloc/assigned once
